FINAM PROJECT/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
aiohttp-retry==2.9.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
spacy==3.7.2
yfinance==0.2.28
numpy==1.24.3
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1
openai==1.3.0